change `AllowOrigin` cors feature to `AspNetCore` cors feature
- update docs and unreleased
