--- src/recipe/Baked.Recipe.Service.Application/Cors/AllowOrigin/AllowOriginCorsExtensions.cs
@@ -1,11 +1,0 @@
-﻿using Baked.Cors;
-using Baked.Cors.AllowOrigin;
-using Baked.Runtime;
-
-namespace Baked;
-
-public static class AllowOriginCorsExtensions
-{
-    public static AllowOriginCorsFeature AllowOrigin(this CorsConfigurator _, params Setting<string>[] origins) =>
-          new(origins);
-}
